+++ NEWS.md
@@ -1,3 +1,9 @@
+# MRFcov 1.0.39
+*   Improvement for the detection of infinite values in data and coordinates
+
+# MRFcov 1.0.38
+*   Option to prep spatial splines separately to allow efficient out-of-sample cross-validation
+
 # MRFcov 1.0.37
 *   More detailed work-throughs of examples in the vignettes
 *   Fixed typo in Bird_Parasite_CRF vignette